backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
